@@ -2,17 +2,17 @@
 from __future__ import annotations
 
 import logging
-from typing import Awaitable, Callable
+from collections.abc import Awaitable, Callable
 
-from aiogram import Router, F
+from aiogram import F, Router
 from aiogram.filters import Command
 from aiogram.types import (
-    Message,
     CallbackQuery,
-    InlineKeyboardMarkup,
     InlineKeyboardButton,
+    InlineKeyboardMarkup,
+    Message,
 )
-from aiohttp import ClientSession, ClientError
+from aiohttp import ClientError, ClientSession
 
 from ..config import settings
 
@@ -26,7 +26,7 @@ class BackendError(Exception):
 
 
 class RateLimitError(Exception):
-    def __init__(self, retry_after: str | None = None):
+    def __init__(self, retry_after: str | None = None) -> None:
         self.retry_after = retry_after
 
 
@@ -61,9 +61,7 @@ async def _request_link_token(chat_id: int) -> tuple[str, str | None]:
 
                 # Логируем тело, чтобы проще было дебажить
                 text = await resp.text()
-                logger.error(
-                    "DFSP /tg/link-start failed: %s %s", resp.status, text
-                )
+                logger.error("DFSP /tg/link-start failed: %s %s", resp.status, text)
                 raise BackendError()
 
     except ClientError as e:
@@ -75,12 +73,7 @@ def _build_link_keyboard(deep_link: str) -> InlineKeyboardMarkup | None:
     if "localhost" in deep_link:
         return None  # не делаем кнопку для локалки
 
-    return InlineKeyboardMarkup(
-        inline_keyboard=[
-            [InlineKeyboardButton(text="🌐 Открыть DFSP", url=deep_link)]
-        ]
-    )
-
+    return InlineKeyboardMarkup(inline_keyboard=[[InlineKeyboardButton(text="🌐 Открыть DFSP", url=deep_link)]])
 
 
 async def _send_link(
@@ -95,21 +88,14 @@ async def _send_link(
             seconds = int(e.retry_after)
 
         if seconds and seconds > 0:
-            text = (
-                "⚠️ Слишком часто запрашиваешь ссылку.\n"
-                f"Попробуй снова примерно через {seconds} секунд."
-            )
+            text = f"⚠️ Слишком часто запрашиваешь ссылку.\nПопробуй снова примерно через {seconds} секунд."
         else:
-            text = (
-                "⚠️ Слишком часто запрашиваешь ссылку.\n"
-                "Попробуй ещё раз чуть позже."
-            )
+            text = "⚠️ Слишком часто запрашиваешь ссылку.\nПопробуй ещё раз чуть позже."
         await send(text, None)
         return
     except BackendError:
         await send(
-            "😔 Сейчас не получается сгенерировать ссылку на привязку.\n"
-            "Попробуй ещё раз чуть позже.",
+            "😔 Сейчас не получается сгенерировать ссылку на привязку.\nПопробуй ещё раз чуть позже.",
             None,
         )
         return
@@ -117,21 +103,25 @@ async def _send_link(
     origin = str(settings.PUBLIC_WEB_ORIGIN).rstrip("/")
     deep_link = f"{origin}/tg/link?token={link_token}"
 
-    text = (
-        "Вот ссылка для привязки аккаунта DFSP к этому Telegram.\n\n"
-        "Ссылка одноразовая и действует ограниченное время. "
-        "Если она истечёт, просто вызови /link ещё раз."
-    )
-    
-    kb = _build_link_keyboard(deep_link)
-    text = (
-        "Вот ссылка для привязки аккаунта DFSP к этому Telegram.\n\n"
-        f"{deep_link}\n\n"
-        "Ссылка одноразовая и действует ограниченное время. "
-        "Если она истечёт, просто вызови /link ещё раз."
-    )
-    await send(text, kb)
+    # Проверка на потенциальные проблемы с конфигурацией
+    from ..utils.diagnostics import check_public_web_origin
 
+    is_valid, error_msg = check_public_web_origin()
+
+    text = (
+        "Вот ссылка для привязки аккаунта DFSP к этому Telegram.\n\n"
+        f"🔗 {deep_link}\n\n"
+        "Ссылка одноразовая и действует ограниченное время. "
+        "Если она истечёт, просто вызови /link ещё раз."
+    )
+
+    if not is_valid:
+        text += f"\n\n⚠️ {error_msg}"
+
+    # Показываем кнопку только для валидного origin
+    kb = _build_link_keyboard(deep_link) if is_valid else None
+
+    await send(text, kb)
 
 
 # --- /link командой ------------------------------------------------------------
@@ -152,16 +142,30 @@ async def cmd_link(message: Message) -> None:
 async def cb_link_start(callback: CallbackQuery) -> None:
     # На всякий случай: если апдейт пришёл не из лички
     if not callback.message:
-        await callback.answer(
-            "Напиши мне в личку, чтобы привязать аккаунт.", show_alert=True
-        )
+        await callback.answer("Напиши мне в личку, чтобы привязать аккаунт.", show_alert=True)
         return
 
     await _send_link(
         chat_id=callback.message.chat.id,
-        send=lambda text, kb: callback.message.answer(
-            text, reply_markup=kb
-        ),
+        send=lambda text, kb: callback.message.answer(text, reply_markup=kb),
     )
     # Закрываем "часики" у пользователя
     await callback.answer()
+
+    # После отправки ссылки показываем инструкцию
+    await callback.message.answer(
+        "📋 <b>Инструкция по привязке:</b>\n\n"
+        "1. Нажми на ссылку выше или скопируй её\n"
+        "2. Открой ссылку в браузере\n"
+        "3. Войди в свой кошелёк\n"
+        "4. Подпиши сообщение для подтверждения\n\n"
+        "После успешной привязки ты получишь уведомление, и сможешь использовать все функции бота!"
+    )
+
+    # Обновляем главное меню
+    from ..handlers import start as start_handlers
+
+    keyboard = start_handlers.get_main_keyboard(is_linked=False)
+    await callback.message.answer(
+        "💡 <b>Главное меню</b>\n\nИспользуй кнопки ниже для навигации:", reply_markup=keyboard
+    )
