@@ -38,34 +38,38 @@ async def cmd_me(message: Message) -> None:
         logger.info("DFSP /bot/me result: %r", profile)
     except Exception:
         logger.exception("Failed to get bot profile from DFSP")
-        await message.answer(
-            "😔 Не удалось получить профиль.\n"
-            "Попробуй ещё раз чуть позже."
-        )
+        await message.answer("😔 Не удалось получить профиль.\nПопробуй ещё раз чуть позже.")
         return
 
     if profile is None:
         # 404 от API — чат не привязан
+        from .start import get_main_keyboard
+
+        keyboard = get_main_keyboard(is_linked=False)
         await message.answer(
-            "К этому чату ещё не привязан кошелёк.\n\n"
+            "❌ К этому чату ещё не привязан кошелёк.\n\n"
             "Чтобы привязать кошелёк:\n"
-            "1. Нажми /link\n"
+            "1. Нажми кнопку «🔗 Привязать аккаунт» ниже\n"
             "2. Открой ссылку в браузере\n"
-            "3. Войди и подпиши сообщение своим кошельком."
+            "3. Войди и подпиши сообщение своим кошельком.",
+            reply_markup=keyboard,
         )
         return
 
     masked = mask_address(profile.address)
     display_name = profile.display_name or "без имени"
 
+    from .start import get_main_keyboard
+
     text = (
-        "👤 *Твой профиль*\n"
-        f"Имя: *{display_name}*\n"
-        f"Адрес: `{masked}`\n\n"
+        "👤 <b>Твой профиль</b>\n\n"
+        f"Имя: <b>{display_name}</b>\n"
+        f"Адрес: <code>{masked}</code>\n\n"
         "Если хочешь отвязать текущий кошелёк — используй команду /unlink.\n"
         "Чтобы привязать другой кошелёк:\n"
         "1. Сначала /unlink\n"
         "2. Затем снова /link и пройди привязку с новым адресом."
     )
 
-    await message.answer(text, parse_mode="Markdown")
+    keyboard = get_main_keyboard(is_linked=True)
+    await message.answer(text, reply_markup=keyboard)
